Document functions across repo
