+++ internal/repository/user_settings.go
@@ -10,6 +10,7 @@ import (
 	"github.com/ilinovom/summary-tasks-bot/internal/model"
 )
 
+// UserSettingsRepository abstracts persistence of user settings.
 type UserSettingsRepository interface {
 	Get(ctx context.Context, userID int64) (*model.UserSettings, error)
 	Save(ctx context.Context, settings *model.UserSettings) error
@@ -24,6 +25,7 @@ type FileUserSettingsRepository struct {
 	data map[int64]*model.UserSettings
 }
 
+// NewFileUserSettingsRepository loads settings from the given JSON file or creates it if missing.
 func NewFileUserSettingsRepository(path string) (*FileUserSettingsRepository, error) {
 	r := &FileUserSettingsRepository{path: path, data: map[int64]*model.UserSettings{}}
 	if err := r.load(); err != nil {
@@ -32,6 +34,7 @@ func NewFileUserSettingsRepository(path string) (*FileUserSettingsRepository, er
 	return r, nil
 }
 
+// load reads the JSON file into memory.
 func (r *FileUserSettingsRepository) load() error {
 	r.mu.Lock()
 	defer r.mu.Unlock()
@@ -47,6 +50,7 @@ func (r *FileUserSettingsRepository) load() error {
 	return json.NewDecoder(file).Decode(&r.data)
 }
 
+// saveLocked writes the in-memory data back to disk.
 func (r *FileUserSettingsRepository) saveLocked() error {
 	file, err := os.Create(r.path)
 	if err != nil {
@@ -58,6 +62,7 @@ func (r *FileUserSettingsRepository) saveLocked() error {
 	return enc.Encode(r.data)
 }
 
+// Get retrieves settings for the user or returns os.ErrNotExist.
 func (r *FileUserSettingsRepository) Get(ctx context.Context, userID int64) (*model.UserSettings, error) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
@@ -68,6 +73,7 @@ func (r *FileUserSettingsRepository) Get(ctx context.Context, userID int64) (*mo
 	return nil, os.ErrNotExist
 }
 
+// Save persists new settings for a user.
 func (r *FileUserSettingsRepository) Save(ctx context.Context, settings *model.UserSettings) error {
 	r.mu.Lock()
 	defer r.mu.Unlock()
@@ -76,6 +82,7 @@ func (r *FileUserSettingsRepository) Save(ctx context.Context, settings *model.U
 	return r.saveLocked()
 }
 
+// Delete removes settings for a user.
 func (r *FileUserSettingsRepository) Delete(ctx context.Context, userID int64) error {
 	r.mu.Lock()
 	defer r.mu.Unlock()
@@ -83,6 +90,7 @@ func (r *FileUserSettingsRepository) Delete(ctx context.Context, userID int64) e
 	return r.saveLocked()
 }
 
+// List returns all stored user settings.
 func (r *FileUserSettingsRepository) List(ctx context.Context) ([]*model.UserSettings, error) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
